extract Paginator component
